requirements.txt
# Spark Core
pyspark>=3.5.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Data Quality
great-expectations>=0.18.0

# Testing
pytest>=7.4.0
pytest-spark>=0.6.0
chispa>=0.9.0

# Linting
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0

# Airflow (for orchestration)
apache-airflow>=2.7.0

# ML
scikit-learn>=1.3.0
